requirements.txt
google-generativeai>=0.4.0
dotenv
